+++ lib/rubocop/cop/util.rb
@@ -153,6 +153,11 @@ module RuboCop
         Parser::Source::Range.new(@processed_source.buffer, begin_pos, end_pos)
       end
 
+      def begins_its_line?(range)
+        source_before_end = range.source_buffer.source[0...range.begin_pos]
+        source_before_end =~ /\n\s*\Z/
+      end
+
       # Returns, for example, a bare `if` node if the given node is an `if`
       # with calls chained to the end of it.
       def first_part_of_call_chain(node)